Change font color in Space 4 and capitalize github
The only remaining problem I see is the size of the anchor (it's bigger
than the button).

I also thought about having Space 1 be the image as the background and
it does look better but I will leave it as it is to make it similar to
TOP required task.

I only see myself modifying this in the future to fix the anchor issue
but from what I have read people would usually just use the following:
<a href="#" class="btn btn-primary" role="button">Get Started</a>
instead of doing an actual button inside an anchor so will keep that
in mind for the future.